Cast enablement check

+++ lib/eventum/class.prefs.php
@@ -37,7 +37,7 @@ class Prefs
             'auto_append_email_sig'         => 'no',
             'auto_append_note_sig'    => 'no',
             'close_popup_windows'     => 0,
-            'relative_date'           => $setup['relative_date'] == 'enabled',
+            'relative_date'           => (int) ($setup['relative_date'] == 'enabled'),
         );
 
         if (is_array($projects)) {
